Docker-build-Added partnerId in request
